Add files via upload
Welcome to Cows and Bulls!

The user needs to guess a random, 4 digit number from 1000 to 9999.  Cows indicate that the user provided an existing digit in its correct position while a bull indicates the user provided an existing digit in the wrong position.  No cows or bulls indicates that the user provided digits that are non existing within the randomly generated number.  The user isn't provided which digits entered are cows and/or bulls and must keep guessing until the full, random, 4 digit number is provided.  The user wins when the random number has been guessed.  The number of times the user guessed is provided upon winning.  Try with your friends and see who can guess the least!